requirements.txt
streamlit
streamlit-extras
python-dotenv
pandas
ccxt
statsmodels